Modification de la mise à l'échelle - suppression des "/5" à chaque coordonnées pour le remplacer par un "*scale" où scale est l'échelle indiquée par le fichier XML

+++ Contrôle Aérien/js/xmlToJavascript.js
@@ -251,7 +251,7 @@ function processXMLData(XMLData) {
 				parseInt(configAttributes["backgroundY"]),
 				parseInt(configAttributes["maxTime"]),
 				configAttributes["performancesTable"],
-				parseInt(configAttributes["scale"]),
+				parseFloat(configAttributes["scale"]),
 				parseInt(configAttributes["windOption"]),
 				parseInt(configAttributes["normVerticalSeparation"]),
 				parseInt(configAttributes["normHorizontalSeparation"]),
